+++ app/src/main/java/ua/hospes/rtm/domain/team/TeamsRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 internal interface TeamsRepository {
     suspend fun get(): List<Team>
 
-    suspend fun get(vararg ids: Int): List<Team>
+    suspend fun get(vararg ids: Long): List<Team>
 
     suspend fun getNotInRace(): List<Team>
 
@@ -15,7 +15,7 @@ internal interface TeamsRepository {
 
     suspend fun save(team: Team)
 
-    suspend fun delete(id: Int)
+    suspend fun delete(id: Long)
 
     suspend fun clear()
 }
